Frontend

+++ frontend/src/components/common/buttons/LikeButton.tsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useState } from "react";
+import { useState, useRef } from "react";
 import useAuthStore from "../../../store/useAuthStore";
 
 interface Like {
@@ -10,12 +10,14 @@ interface LikeButtonProps {
   className?: string;
   postId?: string;
   likes?: Like[];
+  updateLikes?: (newLikes: Like[]) => void;
 }
 
 const LikeButton: React.FC<LikeButtonProps> = ({
   className,
   postId,
   likes,
+  updateLikes,
   ...props
 }) => {
   const user = useAuthStore((state) => state.user);
@@ -24,32 +26,42 @@ const LikeButton: React.FC<LikeButtonProps> = ({
   const [clicked, setClicked] = useState(
     userId && likes?.some((like) => like._id === userId)
   );
+  const lastClickTime = useRef(0);
 
   const handleClick = async () => {
-    if (userId) {
-      try {
-        if (!clicked) {
-          console.log("like");
-          const response = await axios.post(
-            `${process.env.REACT_APP_API_URL}/api/posts/${postId}/likes`,
-            { userId }
-          );
-          console.log(response);
-        } else {
-          console.log("unlike");
-          const response = await axios.delete(
-            `${process.env.REACT_APP_API_URL}/api/posts/${postId}/likes`,
-            {
-              headers: { "user-id": userId },
+    const now = Date.now();
+    if (now - lastClickTime.current >= 1000) {
+      setClicked(!clicked);
+      if (userId) {
+        try {
+          if (!clicked) {
+            console.log("like");
+            const response = await axios.post(
+              `${process.env.REACT_APP_API_URL}/api/posts/${postId}/likes`,
+              { userId }
+            );
+            if (updateLikes) {
+              updateLikes(response.data.data.postLikes);
             }
-          );
-          console.log(response);
+          } else {
+            console.log("unlike");
+            const response = await axios.delete(
+              `${process.env.REACT_APP_API_URL}/api/posts/${postId}/likes`,
+              {
+                headers: { "user-id": userId },
+              }
+            );
+            if (updateLikes) {
+              updateLikes(response.data.data.postLikes);
+            }
+            console.log(response);
+          }
+        } catch (error) {
+          console.error("Error processing the like/unlike action", error);
         }
-      } catch (error) {
-        console.error("Error processing the like/unlike action", error);
       }
+      lastClickTime.current = now;
     }
-    setClicked(!clicked);
   };
 
   const imageUrl = clicked
